Make the code more DRY and more flexible using EventEmitter2. Add also a way to output a log of the `.trigger` actions
The code that was doing event management was copy and pasted in 3 places.
Plus it has a mechanism that prevent the unprocessed listeners to execute if a listener return false. This mechanisms is error prone since it depends on the order in which the event listeners (.on) are registered. It seemed that the mechanism was not used beside optimization.