FIX RTM RETRY AND RECONNECT LOGIC
Existing Botkit slackbot reconnect logic is failing. The retry attributes are passed along with the botkit object but slackbot_worker.js is looking for them in the bot object. Because these attributes do not exist the retry will always fail. There have been several issues opened against this problem and custom code has been written to try to manage the scenario with there is an RTM disconnect and the bot essentially dies for that team because it cannot reconnect.

This fix points to the botkit object in slack_worker.js which then allows the retry logic to work as planned. Since making this change the bot will automatically reconnect and use the reconnect attributes that are used when the controller is initiated.