Store state parameter from oauth against team
I display the 'Add to slack' button on my website and I would like to send the logged in user's id when the button is clicked. To do so I use the `state` parameter but it appears that parameter is currently never accessible. This fix works locally in my certain case, but as I'm new to botkit there may be issues with this implementation that I'm unaware of.