Update core.md re: spawn()
Our team found out that you can simply pass in the string of the team_id that you want to spawn the bot for, which is actually _easier_ than the previous version, but the documentation is super confusing with its reference to dialogContext and turnContext. Hope this or something similar helps.