adds oauth v2 support
Changes to this file are based on an additional Slack adapter option of "oauthVersion". This would allow for oauth v2 support with minimal changes to the adapter. Updates to this file include 2 conditional statements to use new oauth v2 if options.oauthVersion == 'v2', and updates to example oauth response in the comments.