Swap out request for node-fetch
Motivation: `request` is implicitly depended on by
botbuilder-adapter-slack but it's not in package.json.

This means we need to either:
* Add it as a dependency
* Remove it entirely

Given `request` is long since deprecated and usage in
`botbuilder-adapter-slack` is minimal (just this one request), the
latter seems most sensible.

Of candidate HTTP clients to swap out, node-fetch seems the obvious choice:

* It's very small
* It's popular
* some of the botbuilder/botkit packages depend on it already